Fix race conditions in metrics
As detected by the race detector (see
https://travis-ci.org/sky-uk/feed/builds/143821052).
