+++ nginx/nginx_test.go
@@ -20,9 +20,14 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	dto "github.com/prometheus/client_model/go"
 	"github.com/sky-uk/feed/controller"
+	"github.com/sky-uk/feed/util/metrics"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
+
+func init() {
+	metrics.SetConstLabels(make(prometheus.Labels))
+}
 
 const (
 	port          = 9090
